Add files via upload
Plotting incidence of fraud by the job function shows Administrative and Engineering, Full-time and mid-senior  have the highest levels of fraudulent postings. So we should ensure to strictly meet function, employment type and required experience.
Using Random forest, we inferred that “company_logo” is playing a significant role in identifying a job as fake or not. So companies should ensure to have a company logo while posting jobs.
Using Decision Tree by adjusting the threshold we are able to get overall accuracy 83. 03 for identifying for the jobs
After applying  text analytics, result prediction of all models indicated that this problem was high sensitivity and low specificity problem.
We identified different classification machine learning algorithms to see which would result in the highest classification accuracy if a job was fast.